Implement psi4 object factory
