+++ forte/modules/__init__.py
@@ -3,4 +3,5 @@ from .sequential import Sequential
 from .graph_visualizer import GraphVisualizer
 from .options_factory import OptionsFactory
 from .objects_factory_fcidump import ObjectsFactoryFCIDUMP
+from .objects_factory_psi4 import ObjectsFactoryPsi4
 from .mock import HF, FCI, Ints, Ints2, Localizer
