+++ forte/pymodule.py
@@ -51,39 +51,7 @@ from forte.proc.external_active_space_solver import (
     make_hamiltonian,
 )
 from forte.data import ForteData
-from forte.modules import OptionsFactory, ObjectsFactoryFCIDUMP
-
-
-def run_psi4_ref(ref_type, molecule, print_warning=False, **kwargs):
-    """
-    Perform a new Psi4 computation and return a Psi4 Wavefunction object.
-
-    :param ref_type: a Python string for reference type
-    :param molecule: a Psi4 Molecule object on which computation is performed
-    :param print_warning: Boolean for printing warnings on screen
-    :param kwargs: named arguments associated with Psi4
-
-    :return: a Psi4 Wavefunction object from the fresh Psi4 run
-    """
-    ref_type = ref_type.lower().strip()
-
-    if ref_type in ["scf", "hf", "rhf", "rohf", "uhf"]:
-        if print_warning:
-            msg = [
-                "Forte is using orbitals from a Psi4 SCF reference.",
-                "This is not the best for multireference computations.",
-                "To use Psi4 CASSCF orbitals, set REF_TYPE to CASSCF.",
-            ]
-            msg = "\n  ".join(msg)
-            warnings.warn(f"\n  {msg}\n", UserWarning)
-
-        wfn = psi4.driver.scf_helper("forte", molecule=molecule, **kwargs)
-    elif ref_type in ["casscf", "rasscf"]:
-        wfn = psi4.proc.run_detcas(ref_type, molecule=molecule, **kwargs)
-    else:
-        raise ValueError(f"Invalid REF_TYPE: {ref_type.upper()} not available!")
-
-    return wfn
+from forte.modules import OptionsFactory, ObjectsFactoryFCIDUMP, ObjectsFactoryPsi4
 
 
 def check_MO_orthonormality(S, Ca):
@@ -119,316 +87,6 @@ def check_MO_orthonormality(S, Ca):
     else:
         p4print(" Done (OK)\n\n")
         return True
-
-
-def prepare_psi4_ref_wfn(options, **kwargs):
-    """
-    Prepare a Psi4 Wavefunction as reference for Forte.
-    :param options: a ForteOptions object for options
-    :param kwargs: named arguments associated with Psi4
-    :return: (the processed Psi4 Wavefunction, a Forte MOSpaceInfo object)
-
-    Notes:
-        We will create a new Psi4 Wavefunction (wfn_new) if necessary.
-
-        1. For an empty ref_wfn, wfn_new will come from Psi4 SCF or MCSCF.
-
-        2. For a valid ref_wfn, we will test the orbital orthonormality against molecule.
-           If the orbitals from ref_wfn are consistent with the active geometry,
-           wfn_new will simply be a link to ref_wfn.
-           If not, we will rerun a Psi4 SCF and orthogonalize orbitals, where
-           wfn_new comes from this new Psi4 SCF computation.
-    """
-    p4print = psi4.core.print_out
-
-    # grab reference Wavefunction and Molecule from kwargs
-    kwargs = p4util.kwargs_lower(kwargs)
-
-    ref_wfn = kwargs.get("ref_wfn", None)
-
-    molecule = kwargs.pop("molecule", psi4.core.get_active_molecule())
-    point_group = molecule.point_group().symbol()
-
-    # try to read orbitals from file
-    Ca = read_orbitals() if options.get_bool("READ_ORBITALS") else None
-
-    need_orbital_check = True
-    fresh_ref_wfn = True if ref_wfn is None else False
-
-    if ref_wfn is None:
-        ref_type = options.get_str("REF_TYPE")
-        p4print("\n  No reference wave function provided for Forte." f" Computing {ref_type} orbitals using Psi4 ...\n")
-
-        # no warning printing for MCSCF
-        job_type = options.get_str("JOB_TYPE")
-        do_mcscf = job_type in ["CASSCF", "MCSCF_TWO_STEP"] or options.get_bool("CASSCF_REFERENCE")
-
-        # run Psi4 SCF or MCSCF
-        ref_wfn = run_psi4_ref(ref_type, molecule, not do_mcscf, **kwargs)
-
-        need_orbital_check = False if Ca is None else True
-    else:
-        # Ca from file has higher priority than that of ref_wfn
-        Ca = ref_wfn.Ca().clone() if Ca is None else Ca
-
-    # build Forte MOSpaceInfo
-    nmopi = ref_wfn.nmopi()
-    mo_space_info = forte.make_mo_space_info(nmopi, point_group, options)
-
-    # do we need to check MO overlap?
-    if not need_orbital_check:
-        wfn_new = ref_wfn
-    else:
-        # test if input Ca has the correct dimension
-        if Ca.rowdim() != ref_wfn.nsopi() or Ca.coldim() != nmopi:
-            p4print("\n  Expecting orbital dimensions:\n")
-            p4print("\n  row:    ")
-            ref_wfn.nsopi().print_out()
-            p4print("  column: ")
-            nmopi.print_out()
-            p4print("\n  Actual orbital dimensions:\n")
-            p4print("\n  row:    ")
-            Ca.rowdim().print_out()
-            p4print("  column: ")
-            Ca.coldim().print_out()
-            msg = "Invalid orbitals: different basis set / molecule! Check output for more."
-            raise ValueError(msg)
-
-        new_S = psi4.core.Wavefunction.build(molecule, options.get_str("BASIS")).S()
-
-        if check_MO_orthonormality(new_S, Ca):
-            wfn_new = ref_wfn
-            wfn_new.Ca().copy(Ca)
-        else:
-            if fresh_ref_wfn:
-                wfn_new = ref_wfn
-                wfn_new.Ca().copy(ortho_orbs_forte(wfn_new, mo_space_info, Ca))
-            else:
-                p4print("\n  Perform new SCF at current geometry ...\n")
-
-                kwargs_copy = {k: v for k, v in kwargs.items() if k != "ref_wfn"}
-                wfn_new = run_psi4_ref("scf", molecule, False, **kwargs_copy)
-
-                # orthonormalize orbitals
-                wfn_new.Ca().copy(ortho_orbs_forte(wfn_new, mo_space_info, Ca))
-
-                # copy wfn_new to ref_wfn
-                ref_wfn.shallow_copy(wfn_new)
-
-    # set DF and MINAO basis
-    if "DF" in options.get_str("INT_TYPE"):
-        aux_basis = psi4.core.BasisSet.build(
-            molecule,
-            "DF_BASIS_MP2",
-            options.get_str("DF_BASIS_MP2"),
-            "RIFIT",
-            options.get_str("BASIS"),
-            puream=wfn_new.basisset().has_puream(),
-        )
-        wfn_new.set_basisset("DF_BASIS_MP2", aux_basis)
-
-    if options.get_str("MINAO_BASIS"):
-        minao_basis = psi4.core.BasisSet.build(molecule, "MINAO_BASIS", options.get_str("MINAO_BASIS"))
-        wfn_new.set_basisset("MINAO_BASIS", minao_basis)
-
-    return wfn_new, mo_space_info
-
-
-def prepare_forte_objects_from_psi4_wfn(options, wfn, mo_space_info):
-    """
-    Take a psi4 wavefunction object and prepare the ForteIntegrals, SCFInfo, and MOSpaceInfo objects
-
-    Parameters
-    ----------
-    options : ForteOptions
-        A Forte ForteOptions object
-    wfn : psi4 Wavefunction
-        A psi4 Wavefunction object
-    mo_space_info : the MO space info read from options
-        A Forte MOSpaceInfo object
-
-    Returns
-    -------
-    tuple(ForteIntegrals, SCFInfo, MOSpaceInfo)
-        a tuple containing the ForteIntegrals, SCFInfo, and MOSpaceInfo objects
-    """
-
-    # Call methods that project the orbitals (AVAS, embedding)
-    mo_space_info = orbital_projection(wfn, options, mo_space_info)
-
-    # Build Forte SCFInfo object
-    scf_info = forte.SCFInfo(wfn)
-
-    # Build a map from Forte StateInfo to the weights
-    state_weights_map = forte.make_state_weights_map(options, mo_space_info)
-
-    return (state_weights_map, mo_space_info, scf_info)
-
-
-# def make_state_info_from_fcidump(fcidump, options):
-#     nel = fcidump["nelec"]
-#     if not options.is_none("NEL"):
-#         nel = options.get_int("NEL")
-
-#     multiplicity = fcidump["ms2"] + 1
-#     if not options.is_none("MULTIPLICITY"):
-#         multiplicity = options.get_int("MULTIPLICITY")
-
-#     # If the user did not specify ms determine the value from the input or
-#     # take the lowest value consistent with the value of "MULTIPLICITY"
-#     # For example:
-#     #    singlet: multiplicity = 1 -> twice_ms = 0 (ms = 0)
-#     #    doublet: multiplicity = 2 -> twice_ms = 1 (ms = 1/2)
-#     #    triplet: multiplicity = 3 -> twice_ms = 0 (ms = 0)
-#     twice_ms = (multiplicity + 1) % 2
-#     if not options.is_none("MS"):
-#         twice_ms = int(round(2.0 * options.get_double("MS")))
-
-#     if ((nel - twice_ms) % 2) != 0:
-#         raise Exception(f"Forte: the value of MS ({twice_ms}/2) is incompatible with the number of electrons ({nel})")
-
-#     na = (nel + twice_ms) // 2
-#     nb = nel - na
-
-#     irrep = fcidump["isym"]
-#     if not options.is_none("ROOT_SYM"):
-#         irrep = options.get_int("ROOT_SYM")
-
-#     return forte.StateInfo(na, nb, multiplicity, twice_ms, irrep)
-
-
-def prepare_forte_objects_from_fcidump(options, path="."):
-    fcidump_file = options.get_str("FCIDUMP_FILE")
-    filename = pathlib.Path(path) / fcidump_file
-    psi4.core.print_out(f"\n  Reading integral information from FCIDUMP file {filename}")
-    fcidump = forte.proc.fcidump_from_file(filename, convert_to_psi4=True)
-
-    irrep_size = {"c1": 1, "ci": 2, "c2": 2, "cs": 2, "d2": 4, "c2v": 4, "c2h": 4, "d2h": 8}
-
-    nmo = len(fcidump["orbsym"])
-    if "pntgrp" in fcidump:
-        nirrep = irrep_size[fcidump["pntgrp"].lower()]
-        nmopi_list = [fcidump["orbsym"].count(h) for h in range(nirrep)]
-    else:
-        fcidump["pntgrp"] = "C1"  # set the point group to C1
-        fcidump["isym"] = 0  # shift by -1
-        nirrep = 1
-        nmopi_list = [nmo]
-
-    nmopi_offset = [sum(nmopi_list[0:h]) for h in range(nirrep)]
-
-    nmopi = psi4.core.Dimension(nmopi_list)
-
-    # Create the MOSpaceInfo object
-    mo_space_info = forte.make_mo_space_info(nmopi, fcidump["pntgrp"], options)
-
-    # Call methods that project the orbitals (AVAS, embedding)
-    # skipped due to lack of functionality
-
-    # manufacture a SCFInfo object from the FCIDUMP file (this assumes C1 symmetry)
-    nel = fcidump["nelec"]
-    ms2 = fcidump["ms2"]
-    na = (nel + ms2) // 2
-    nb = nel - na
-    if fcidump["pntgrp"] == "C1":
-        doccpi = psi4.core.Dimension([nb])
-        soccpi = psi4.core.Dimension([ms2])
-    else:
-        doccpi = options.get_int_list("FCIDUMP_DOCC")
-        soccpi = options.get_int_list("FCIDUMP_SOCC")
-        if len(doccpi) + len(soccpi) == 0:
-            print("Reading a FCIDUMP file that uses symmetry but no DOCC and SOCC is specified.")
-            print("Use the FCIDUMP_DOCC and FCIDUMP_SOCC options to specify the number of occupied orbitals per irrep.")
-            doccpi = psi4.core.Dimension([0] * nirrep)
-            soccpi = psi4.core.Dimension([0] * nirrep)
-
-    if "epsilon" in fcidump:
-        epsilon_a = psi4.core.Vector.from_array(fcidump["epsilon"])
-        epsilon_b = psi4.core.Vector.from_array(fcidump["epsilon"])
-    else:
-        # manufacture Fock matrices
-        epsilon_a = psi4.core.Vector(nmo)
-        epsilon_b = psi4.core.Vector(nmo)
-        hcore = fcidump["hcore"]
-        eri = fcidump["eri"]
-        nmo = fcidump["norb"]
-        for i in range(nmo):
-            val = hcore[i, i]
-            for h in range(nirrep):
-                for j in range(nmopi_offset[h], nmopi_offset[h] + doccpi[h] + soccpi[h]):
-                    val += eri[i, i, j, j] - eri[i, j, i, j]
-                for j in range(nmopi_offset[h], nmopi_offset[h] + doccpi[h]):
-                    val += eri[i, i, j, j]
-            epsilon_a.set(i, val)
-
-            val = hcore[i, i]
-            for h in range(nirrep):
-                for j in range(nmopi_offset[h], nmopi_offset[h] + doccpi[h] + soccpi[h]):
-                    val += eri[i, i, j, j]
-                for j in range(nmopi_offset[h], nmopi_offset[h] + doccpi[h]):
-                    val += eri[i, i, j, j] - eri[i, j, i, j]
-            epsilon_b.set(i, val)
-
-    scf_info = forte.SCFInfo(nmopi, doccpi, soccpi, 0.0, epsilon_a, epsilon_b)
-
-    state_info = make_state_info_from_fcidump(fcidump, options)
-    state_weights_map = {state_info: [1.0]}
-    return (state_weights_map, mo_space_info, scf_info, fcidump)
-
-
-# def make_ints_from_fcidump(fcidump, data: ForteData):
-#     # transform two-electron integrals from chemist to physicist notation
-#     eri = fcidump["eri"]
-#     nmo = fcidump["norb"]
-#     eri_aa = np.zeros((nmo, nmo, nmo, nmo))
-#     eri_ab = np.zeros((nmo, nmo, nmo, nmo))
-#     eri_bb = np.zeros((nmo, nmo, nmo, nmo))
-#     # <ij||kl> = (ik|jl) - (il|jk)
-#     eri_aa += np.einsum("ikjl->ijkl", eri)
-#     eri_aa -= np.einsum("iljk->ijkl", eri)
-#     # <ij|kl> = (ik|jl)
-#     eri_ab = np.einsum("ikjl->ijkl", eri)
-#     # <ij||kl> = (ik|jl) - (il|jk)
-#     eri_bb += np.einsum("ikjl->ijkl", eri)
-#     eri_bb -= np.einsum("iljk->ijkl", eri)
-
-#     ints = forte.make_custom_ints(
-#         data.options,
-#         data.mo_space_info,
-#         fcidump["enuc"],
-#         fcidump["hcore"].flatten(),
-#         fcidump["hcore"].flatten(),
-#         eri_aa.flatten(),
-#         eri_ab.flatten(),
-#         eri_bb.flatten(),
-#     )
-#     data.ints = ints
-#     return data
-
-
-def prepare_forte_objects(data, name, **kwargs):
-    """
-    Prepare the ForteIntegrals, SCFInfo, and MOSpaceInfo objects.
-    :param data: the ForteOptions object
-    :param name: the name of the module associated with Psi4
-    :param kwargs: named arguments associated with Psi4
-    :return: a tuple of (Wavefunction, ForteIntegrals, SCFInfo, MOSpaceInfo, FCIDUMP)
-    """
-    lowername = name.lower().strip()
-    options = data.options
-
-    psi4.core.print_out("\n\n  Preparing forte objects from a Psi4 Wavefunction object")
-    ref_wfn, mo_space_info = prepare_psi4_ref_wfn(options, **kwargs)
-    forte_objects = prepare_forte_objects_from_psi4_wfn(options, ref_wfn, mo_space_info)
-    state_weights_map, mo_space_info, scf_info = forte_objects
-    fcidump = None
-
-    data.mo_space_info = mo_space_info
-    data.scf_info = scf_info
-    data.state_weights_map = state_weights_map
-    data.psi_wfn = ref_wfn
-
-    return data, fcidump
 
 
 def forte_driver(data: ForteData):
@@ -517,7 +175,6 @@ def run_forte(name, **kwargs):
     >>> energy('forte')
 
     """
-
     # # Start Forte, initialize ambit
     # my_proc_n_nodes = forte.startup()
     # my_proc, n_nodes = my_proc_n_nodes
@@ -528,6 +185,8 @@ def run_forte(name, **kwargs):
     # Print the banner
     forte.banner()
 
+    job_type = data.options.get_str("JOB_TYPE")
+
     # Start timer
     start_pre_ints = time.time()
 
@@ -536,29 +195,7 @@ def run_forte(name, **kwargs):
         psi4.core.print_out("\n  Forte will use custom integrals")
         data = ObjectsFactoryFCIDUMP(options=kwargs).run(data)
     else:
-        data, fcidump = prepare_forte_objects(data, name, **kwargs)
-
-    job_type = data.options.get_str("JOB_TYPE")
-    if job_type == "NONE" and data.options.get_str("ORBITAL_TYPE") == "CANONICAL":
-        psi4.core.set_scalar_variable("CURRENT ENERGY", 0.0)
-        return data.psi_wfn
-
-    # these two functions are used by the external solver to read and write MO coefficients
-    if data.options.get_bool("WRITE_WFN"):
-        write_wavefunction(data)
-
-    if data.options.get_bool("READ_WFN"):
-        if not os.path.isfile("coeff.json"):
-            print("No coefficient files in input folder, run a SCF first!")
-            exit()
-        read_wavefunction(data)
-
-    if "FCIDUMP" in data.options.get_str("INT_TYPE"):
-        pass
-    else:
-        psi4.core.print_out("\n  Forte will use psi4 integrals")
-        # Make an integral object from the psi4 wavefunction object
-        data.ints = forte.make_ints_from_psi4(data.psi_wfn, data.options, data.mo_space_info)
+        data = ObjectsFactoryPsi4(options=kwargs).run(data)
 
     start = time.time()
 
@@ -569,7 +206,7 @@ def run_forte(name, **kwargs):
         orb_t.compute_transformation()
         Ua = orb_t.get_Ua()
         Ub = orb_t.get_Ub()
-        ints.rotate_orbitals(Ua, Ub, job_type != "NONE")
+        data.ints.rotate_orbitals(Ua, Ub, job_type != "NONE")
 
     # Run a method
     if job_type == "NONE":
